+++ crates/tools/tool-whisper/src/chat_session.rs
@@ -1,18 +1,19 @@
-use crate::connection::Connection;
+use crate::connection::TcpConnection;
 use crate::encrypt::encryption::Encryption;
 use crate::encrypt::message_decrypter::MessageDecrypter;
 use crate::encrypt::message_encrypter::MessageEncrypter;
+use crate::models::shared_types::UiMessage;
 use anyhow::Result;
 use std::sync::mpsc::{Receiver, Sender};
 
 pub struct ChatSession {
     role: String,
-    connection: Connection,
+    connection: TcpConnection,
     my_encryption: Encryption,
     peer_encryption: MessageEncrypter,
     tx_outgoing_messages: Sender<String>,
-    tx_incoming_messages: Sender<String>,
-    pub rx_incoming_messages: Receiver<String>,
+    tx_incoming_messages: Sender<UiMessage>,
+    pub rx_incoming_messages: Receiver<UiMessage>,
     rx_outgoing_messages: Receiver<String>,
 }
 
@@ -20,12 +21,12 @@ impl ChatSession {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         role: String,
-        connection: Connection,
+        connection: TcpConnection,
         my_encryption: Encryption,
         peer_encryption: MessageEncrypter,
         tx_outgoing_messages: Sender<String>,
-        tx_incoming_messages: Sender<String>,
-        rx_incoming_messages: Receiver<String>,
+        tx_incoming_messages: Sender<UiMessage>,
+        rx_incoming_messages: Receiver<UiMessage>,
         rx_outgoing_messages: Receiver<String>,
     ) -> Self {
         Self {
@@ -58,7 +59,7 @@ impl ChatSession {
         std::mem::replace(&mut self.rx_outgoing_messages, dummy_rx)
     }
 
-    pub fn get_incoming_transmitter(&self) -> Sender<String> {
+    pub fn get_incoming_transmitter(&self) -> Sender<UiMessage> {
         self.tx_incoming_messages.clone()
     }
 
@@ -66,14 +67,14 @@ impl ChatSession {
         self.tx_outgoing_messages.clone()
     }
 
-    pub fn take_incoming_receiver(&mut self) -> Receiver<String> {
+    pub fn take_incoming_receiver(&mut self) -> Receiver<UiMessage> {
         // We need to replace the receiver with a dummy one since we're moving it out
         let (_, dummy_rx) = std::sync::mpsc::channel();
         std::mem::replace(&mut self.rx_incoming_messages, dummy_rx)
     }
 
-    pub fn split_connection(&mut self) -> Result<Connection> {
+    pub fn split_connection(&mut self) -> Result<TcpConnection> {
         let conn = self.connection.connection.try_clone()?;
-        Connection::new_from_connection(conn)
+        TcpConnection::new_from_connection(conn)
     }
 }