+++ crates/tools/tool-whisper/src/encrypt/mod.rs
@@ -1,4 +1,3 @@
-pub mod encrypt_traits;
 pub mod encryption;
 pub mod message_decrypter;
 pub mod message_encrypter;